use a dedicated AttributePreFiltering class
Given that the existing attributing processing methods and the new
attribute pre-filtering methods don't share any content or logic, have
the pre-filtering methods live in a separate dedicated
`AttributePreFiltering` class.
